+++ server/app/entity/user.go
@@ -1,7 +1,7 @@
 package entity
 
 type User struct {
-	ID          uint64
+	ID          uint32
 	Username    string
 	AvatarURL   string
 	CurrentRank uint16
