+++ server/app/internal/redis/multiplayer_lobby.go
@@ -9,7 +9,7 @@ import (
 	"github.com/go-redis/redis"
 )
 
-func (r *RedisClient) JoinLobby(ctx context.Context, lobbyID uint64, playerID uint64, rank uint16) error {
+func (r *RedisClient) JoinLobby(ctx context.Context, lobbyID uint32, playerID uint32, rank uint16) error {
 	key := fmt.Sprintf("Lobby:%d", lobbyID)
 	_, err := r.redis_conn.ZAdd(key, redis.Z{
 		Score:  float64(rank),
